requirements.txt
ctransformers==0.2.5
sentence-transformers==2.6.0
pinecone-client==3.0.0
langchain==0.2.2
langchain-huggingface==0.0.3
flask
pypdf==3.17.0
python-dotenv
-e .
